0.6.4 Changes 1
Fixes:
Fixed a bug when you can remove a core system with central black hole in detailed generation.

New:

Changes:
Scrolling is now dynamic. (There is small visual bug when scroll out but it doesn't affect anything, ignore that,. I'll fix it a one day)
Map name is now shown and can be changed on click (Click name to change it), default is MultiGalaxy.